Add anti-CSRF to confirm modals
CKAN has confirmation modals for certain actions such as deleting a
resource or a package. Since these are done in javascript, the form is
built on the fly. These forms thus did not inherit the CSRF tokens.

This commit adds the CSRF tokens to the javascript-built forms by
passing them as an HTML attribute on the action buttons.

Additionally, CSRF fixes were made to the package and user controller.